refactor: actor model
Evolving the "chan_rpc" module into what it really is: an actor model.

Several frameworks were evaluated, but the problem with all of them is that
they're frameworks.

This is basically renaming things; a "Service" is really an Actor.

tx and rx are encapsulated in ChanClient and ChanServer types. ChanClient is
further wrapped in an Operator which serves as a handle for interacting with
the actor.

This should hopefully make these actors easier to coordinate at the top.
Mucking around with tasks and chans was too tedious there.