refactor: introduce runners, improve operator & chan abstractions, errors
Operators now take a runner as an argument, which provides a strategy for
dealing with errors in operated actors.

Because actors are re-run by runners, their channels need to be cloned. It
turns out that mpsc is ill-suited to this task, so mpsc is replaced with flume.

Wrapping anyhow::Error with a struct also turned out to be ill-advised,
conflicting quite a bit with anyhow's From trait impl. So helper methods were
moved to functions in the error module, but all errors are aliased
anyhow::Error.

Renamed "peer" to "node" because it's a veilid node. Peer is now reserved for
the supervisor that will coordinate all the operated actors, coming in a
followup.

Added an example share_announcer, which demonstrates announcing a share to
Veilid DHT under an operator with a runner that holds execution until the
veilid network is available.