refactor!: indicate block not found with none
Use Option<&[u8]> to indicate presence of a block (Some) or lack thereof
(None).

Protocol remains the same, but the Node API is updated to make this distinction
clear.

This does have the consequence that there's no way to positively acknowledge an
empty block with something like Some(&[]), which should be fine.