refactor!: replace send/recv pattern with call/defer
Following Alice's advice on bespoke actors a little more closely with this one.

Requests embed an optional oneshot response, which is used to provide the
response to a specific request.

Adding a defer which mates the oneshot with a flume channel, for request
queuing & streaming use cases.

This is a pretty drastic refactoring, but it's all still unreleased, so it'll
go into the same 0.5 in progress.