fix: deal with unrecoverable errors intentionally
Channel send and receive errors are generally unrecoverable in stigmerge. If
they happen, it means there's a serious problem where an actor or its
subscription was dropped in an unexpected way.

The fetcher has a few other unrecoverable errors; if it can't access the local
filesystem to diff the have vs want share, the application cannot recover and
should shut down on error.

Generally any fetcher state handling method should return an error only if it
needs to shut down. So transient errors within the fetcher, such as a failure
to select viable peers, should not be returned from the fetch loop.