fix: separate fetch and verify status
Fetching blocks and verifying pieces are handled by distinct async actors. The status
for fetching and verifying was in contention during a fetch, when they should be
represented separately, if both are going to be represented.

For now, we'll show verification as a separate progress bar that clears once
fetch is done.