refactor: extract high level share api
Encapsulate the peer setup necessary to operate a single payload share.

This basically provides a high-level API to applications that want to manage
one or more shares concurrently in a user agent, whether a simple single-share
CLI, a TUI that manages multiple shares, or a multi-platform GUI.

This refactor replatforms the existing CLI to use the new high-level API.

Satisfies #326.