fix: join share announcer with other share operators
The share announcer wasn't being joined with the stigmerge::share::Share join
set.  It was dropped after starting the share. This allows the share to
announce once, but doesn't reannounce on connection changes and
route changes.

This may fix the issue identified in #347 without the need for a regular ping.