feat: expose veilid config in crate api
Add a routing_context constructor that accepts a VeilidConfig, building on the
config-struct based approach.

Removed the NODE_ADDR env var; for a seedbox that wants to expose listeners,
we'll make that decision up the stack (in the stigmerge CLI, UI, etc).

This also delegates the decision whether to UPnP or not, which can be
controversial. I don't allow UPnP on my home network, but surprise UPnP might
be unpleasant for some.