fix: recover lagged broadcast receiver
In the event a broadcast receive errors, it could be due to lag. Larger buffers
can help avoid lag, but some lag may be inevitable due to async scheduling --
imagine a Maxwell's demon that starves the receiver long enough for its buffer
to fill.

This is a brute-force patch over the problem; it would be better to make
auto-recovery endemic somehow to how all broadcast channels are used, but first
let's see if this fixes long-term process instability.